Add check for compressed kernel
The kernel passed to vfkit must be uncompressed, or the VM will silently
fail to start hanging at a black screen.
This commit checks if the kernel is compressed or not before attempting
to start the VM, and errors out if this is the case.

README.md has more details why an uncompressed kernel is required since
commit c92dd0416.