build: Build for macOS11+
Since the macOS12-only API in github.com/Code-Hex/vz are annotated
with `if (@available(macOS 12, *)) { }`, an error will be safely
returned if one attempts to use them on macOS11.
This commit adds -mmacosx-version-min=11.0 to ensure the built binary
will be runnable on macOS 11 regardless of the OS it's built on.