timesync: Synchronize guest time after suspend
This uses prashantgupta24/mac-sleep-notifier
This makes use of cgo, but vfkit must already be built on a macOS
machine, so this does not add a new requirement.

mac-sleep-notifier sends a message over a go channel when a sleep/wakeup
event occurs. vfkit catches it, and sends a 'guest-set-time' command to
the qemu-guest-agent service running inside the guest.

This ensures the host/guest time are synchronized after coming out of
sleep.