Ignore SIGPIPE
vfkit started by crc was exiting after a few suspend/resume cycles.
This seems to have been caused by a SIGPIPE (exit status 13, and this
was also reported by lldb).

golang sometimes exits the current process upon receiving a SIGPIPE:
https://pkg.go.dev/os/signal#hdr-SIGPIPE

This commit catches and ignores SIGPIPE to workaround this crc issue.

Ideally we'll take a closer why this happens in the future.