cmdline: Add stringSlice type implementing pflag.Value
pflag parses string slices uses encoding/csv. This has one annoying
limitation, it does not handle '--arg foo="a,b"', it returns an error
about quote issues.
Since I'd like to use a syntax like:
--bootloader linux,cmdline="quiet console=ttyS0,115200"
in vfkit, this commit is adding a string slice implementation which does
not have this limitation when parsing.
One thing which is not handled is \" to have quotes inside quotes.