build: Add entitlement to amd64/arm64 binaries
It's not possible to build a universal cgo binary in brew, the build
fails when trying to compile a go program for the non-native arch.
We'll only build a binary for the current arch on brew. These binaries
need to have the virtualization entitlement, so instead of adding the
entitlement to the final universal binary, we add it instead to each
arch-specific binary.