Report errors from the virtualization framework
When the VM is running, the virtualization framework can report errors
when the virtualization stack failed for any reasons (compressed kernel
on M1 for example).
waitForVMState report these errors, but vfkit's "main loop" assumes all
errors to be timeouts from waitForVMState.
This commit makes the check more finegrained so that vfkit exits on
virtualization errors.