Add Restful service
add restful service to VFKit which defaults to `tcp://localhost:8081`.  The URL can be changed using the newly introduced --restful-uri
option.  The schemes may be `tcp` or `unix` (unix domain socket).  The service for the unix domain implementation is incomplete and
should be considered a TODO still.  users can also opt out of the service with a `--restful-uri none`.

the restful service now replies to three endpoints:

/vm/state GET which returns the state of the vm
/vm/state POST which allows for a state change
/vm/inspect GET which returns basic information about the virtual machine

error handling for the endpoints needs to be completed still.  these endpoints should return http response codes (like 200) based on
the error handling.  this can be completed in a subsequent PR.

/version GET should also probably be added. I was unsure if we should return the version of vfkit (my first inclination) or some
sort of API version.  maybe we should return both?

added tests for new content where applicable

Signed-off-by: Brent Baude <bbaude@redhat.com>