custom json methods for virtual machine
Add custom JSON methods for Marshal and Unmarshal so other projects can write a configuration file based on virtual machine.  the key issue is the use of interfaces, for example devices, in virtual machine.  When using JSON to marshal and unmarshal, we need to tag what kind of device it is so the process can tag and ultimately re-identify the proper device.

Signed-off-by: Brent Baude <bbaude@redhat.com>