serial: Add --device virtio-serial,pty support
This new virtio-serial option allocates a pseudo-tty for the VM console.
It can then be accessed using `screen` for example.
This is a bit similar to the `--device virtio-serial,stdio` option,
except that the console is not tied to the terminal running vfkit, it's
possible to connect/disconnect from the pseudo-tty from any terminal.

This fixes https://github.com/crc-org/vfkit/issues/48

Signed-off-by: Christophe Fergeau <cfergeau@redhat.com>