bootloader: Rework `isUncompressedKernel`
So far we were using `h2non/filetype` which is not a great fit for what
we want to do, as kernels do not have their own mimetypes, and there is
no nice way to convey compressed/uncompressed/arch through mime types.

This was working through heuristics so far, but the used heuristics
break for the kernel used by puipui linux
(https://github.com/Code-Hex/puipui-linux)

This commit removes the use of h2non/filetype, and reuses the patterns
from https://github.com/file/file/blob/master/magic/Magdir/linux to do
the detection we need.

This only adds a magic for uncompressed arm64 kernels.
There's also a pattern for compressed kernels, but it does not match the arm64
kernels I tested with. They are either detected as gzip files, or as EFI files,
and don't contain the magic pattern.