feat(rest): support check status can be change
In the previous API, there was no way to detect before changing the
status, which would make it difficult for developers to handle the
following situations:

    - When the virtual machine status is Stop, calling Pause may fail.
    - When the virtual machine status is Pause, setting the status to
      Pause again may fail.

In the previous solution, developers needed to first get the current
status of the virtual machine and then decide whether to proceed to the
next step. However, the current virtual machine status is very diverse
(see:
https://github.com/Code-Hex/vz/blob/bd29a7ea3d39465c4224bfb01e990e8c220a8449/virtualization.go#L23),
which would require developers to handle various cases perfectly. This
PR solves this problem.