version: Add moduleVersionFromBuildInfo
This adds correct version information for `vfkit -version` when
it's installed using:
```
go install github.com/crc-org/vfkit/cmd/vfkit
```

However, due to missing entitlements, a binary installed this way is
unlikely to be able to start VMs.
I'm keeping this codepath to use the same code in gvisor-tap-vsock and
vfkit.

Signed-off-by: Christophe Fergeau <cfergeau@redhat.com>