refactor: improve code for wait vm to reach the specified state
In previous implementations, there was a need to maintain a timeout error variable to ignore while waiting for the VM to stop.
This part of the code was not intuitive enough, and other maintainers couldn’t grasp its intention at first glance.
This improvement has optimized the `waitForVMState` function by adding a timeout parameter, making the code more intuitive.

Signed-off-by: Black-Hole1 <bh@bugs.cc>