ghactions: Adjust action so that versioning works
For `git describe` to work on non-tag pushes, we need the repo
history/tags, a shallow clone with just the latest commit won't be
enough.

There's also a bug in actions/checkout for tag pushes, the tag
is overridden with a non-tag, which breaks `git describe` for the tag.
This commit adds a workaround for this.

Signed-off-by: Christophe Fergeau <cfergeau@redhat.com>