rest: Report memory/cpu/devices
This currently is only a stub with hardcoded values.
This commit adds a reference to
config.VirtualMachine to VzVirtualMachine. We can then serialize it to
JSON as there is already code for this.

This changes the Memory field to MemoryBytes, but since it contained a
dummy value so far, this should not be a big problem

This is based on a patch from Kevin Cui <bh@bugs.cc>

Signed-off-by: Christophe Fergeau <cfergeau@redhat.com>