test: Catch vfkit process exiting unexpectedly
This commit an a go routine which will `Wait()` for the vfkit process.
When it exits, it will push the corresponding error (or nil) to a
channel.
The code waiting for the VM to startup can then check for errors in
their select{} retry loop, and fail acccordingly.

This will make debugging test failures easier, as currently if the vfkit
process dies early, the test appears to be taking a very long time
without any indication as to what's going on.

Signed-off-by: Christophe Fergeau <cfergeau@redhat.com>