rest: Check UNIX socket path length
UNIX socket paths have a 104 bytes upper limit which is not so hard to
trigger with the temporary directories created during testing:
`/var/folders/l0/rh4v8_j54k37h2w320__7d1h0000gn/T/TestVsockConnect2047157905/001/`

Better to detect this early and return an explicit error when this
happens.

This also modifies the REST API test, it shortens one of the test names
and the rest socket name as these were triggering this problem.

Signed-off-by: Christophe Fergeau <cfergeau@redhat.com>