vf: Add VirtualMachineConfig type
This commit renames vzVirtualMachineConfiguration to
VirtualMachineConfig.
vzVirtualMachineConfiguration was a wrapper over the Code-Hex/vz type
vz.VirtualMachineConfiguration.

VirtualMachineConfig improves vzVirtualMachineConfiguration because it
also keeps around the config.VirtualMachine go object which was used to
created the vz.VirtualMachineConfiguration object. This will be useful
in simplifying the code using the pkg/vf package.

Signed-off-by: Christophe Fergeau <cfergeau@redhat.com>