vf: Add vf.VirtualMachine type
Similarly to the vf.VirtualMachineConfiguration type introduced in the previous
commit, the vf.VirtualMachine type wraps both a vz.VirtualMachine
instance, and the vfkit/go data which was used to generate it, which is
the aforementioned VirtualMachineConfiguration type.

This allows to simplify some users of pkg/vf, in particular the rest
code. `code-hex/vz` is now only imported in pkg/vf

Signed-off-by: Christophe Fergeau <cfergeau@redhat.com>