json: test: Add 'stability' JSON test
We don't want the JSON serialization of pkg/config/ data structures to
change between releases.
This commit makes use of the `reflect` package to test the JSON serialization
of as many fields as possible. If one of these field change, the test
will make us aware of it.
We still need to explicitly list the types to be tested (ie the types
which can be serialized to JSON).

Signed-off-by: Christophe Fergeau <cfergeau@redhat.com>