rest: Refactor/simplify vf code
The rest/vf package introduces a `VzVirtualMachine` type to wrap virtual
machine interactions. This type is a struct with a
`VzVM *vz.VirtualMachine` member.

This commit makes this struct member anonymous instead of naming it
`VzVM`, this allows to remove the VzVirtualMachine state related methods
which were thin wrappers over methods with (almost) the same name
already defined in `*vz.VirtualMachine`.

Signed-off-by: Christophe Fergeau <cfergeau@redhat.com>