gha: Use 'make lint' instead of golangci-lint-action
While golangci/golangci-lint-action is convenient to use, it has the
side-effect of always using the latest golangci-lint release, which can
introduce new failures at any time.
If we use `make lint`, the golangci-lint version we tested with will be
used, and we can control when to updated to newer golangci-lint version,
and address the new warnings.
For example, golangci-lint v1.60.2 introduces new checks for int
overflows, which I'd rather address at a later time.

Signed-off-by: Christophe Fergeau <cfergeau@redhat.com>