doc: Set minimum supported version to macOS 12
macOS 11 github runners are no longer available so we no longer build
and test vfkit on this platform.
macOS 11 has been out of support for a year now.
This commit changes the minimum supported version to macOS 12.
A mention of macOS 11 is kept, we'll remove it when we'll know for sure
vfkit is broken on this platform

Signed-off-by: Christophe Fergeau <cfergeau@redhat.com>