test: add e2e test with fedora
it adds an e2e test which downloads fedora 40 and create a VM using it. Then mounts an ISO image containing the user-data and meta-data files needed by cloud-init to apply our configuration.
The meta-data file is empty but the user-data file contains info about a new user that has to be added to the system.
After the VM gets started it verifies that the user has been actually added by connecting using SSH.
At the end, the VM is stopped.