ignition: Start the http server directly over vsock
Currently, the ignition http server listens on a unix socket. The
network traffic is then tunneled over vsock.
However, as the http server runs in the vfkit process, we can
directly create a vsock listener to communicate with the VM.
This should be more efficient as this removes the tcpproxy layer, and
more importantly, this removes the creation of the on-disk unix socket,
which had a hardcoded name.

This fixes https://github.com/crc-org/vfkit/issues/369

Signed-off-by: Christophe Fergeau <cfergeau@redhat.com>