json: Add backward compatibility with VfkitMagic
The previous commits added a VfkitMagic property which for now defaults
to true for backwards compatibility with older gvproxy versions.
However, this default was not used when unmarshalling a json vm
configuration made with an older vfkit version. This could cause
unexpected issues with networking.

This commit sets VfkitMagic to true by default when unmarshalling json.

Signed-off-by: Christophe Fergeau <cfergeau@redhat.com>
Assisted-by: CodeRabbit.ai
Signed-off-by: Christophe Fergeau <cfergeau@redhat.com>