virtio-net: Close also fd= on shutdown
Previously we closed the fd obtained from connecting to unixSocketPath=
option on shutdown, but we did not close the fd obtained from the fd=
option. This change ensures that we close the fd in both cases. This
can help the network proxy to detect vm shutdown.

In practice this does not help the network proxy to detect shutdown
since datagram socket is connectionless, and read does not return when
the peer closes the socket. This change is only for consistency with
unixSocketPath.