virtio-net: Fix fd close for unixSocketPath= option
To pass the fd to VZ, we use conn.File(). This duplicate the connection
file descriptor and create an independent file descriptor. Closing the
connection does not effect the duplicate descriptor, and closing the
duplicate descriptor does not affect the connection.

Since we pass the duplicate descriptor to VZ we should close the
connection. This ensure that the network proxy can detect when we close
out duplicate descriptor on shutdown.

In practice this does not help the network proxy to detect shutdown
since datagram socket is connectionless and read() does not return when
the peer closes the socket. This change is only makes the intent more
clear.