virtio-net: Fix Shutdown and improve logging
Previously if removing the local socket failed, we would not close the
socket and return an error that would be ignored silently.

In shutdown flow we must perform all operations so we cannot fail on the
first error. Change Shutdown() to return nothing and log the action it
takes and any errors.

With this change we can tell if the shutdown function was running by
inspecting the debug logs, and shutdown errors will be logged.

Example debug logs on shutdown with unixSocketPath=:

    2025/10/12 18:30:20 captured terminated, calling exit handlers and exiting..
    time="2025-10-12T18:30:20+03:00" level=debug msg="shutting down..."
    time="2025-10-12T18:30:21+03:00" level=info msg="VM is stopped"
    time="2025-10-12T18:30:21+03:00" level=debug msg="Removing /Users/nir/.vmnet-helper/vms/vfkit/vfkit-a52a-3bbd.sock"
    time="2025-10-12T18:30:21+03:00" level=debug msg="Closing fd 8"

Example debug logs on shutdown with fd=:

    2025/10/12 18:35:25 captured terminated, calling exit handlers and exiting..
    time="2025-10-12T18:35:25+03:00" level=debug msg="shutting down..."
    time="2025-10-12T18:35:26+03:00" level=info msg="VM is stopped"
    time="2025-10-12T18:35:26+03:00" level=debug msg="Closing fd 3"